requirements.txt
fastapi
uvicorn
typing
supabase
python-dotenv
pika
pydantic
pymongo